feat: 重构代码结构，将json和json5分离
