+++ DanKeJson/Utils/CommentParser.cs
@@ -1,28 +1,8 @@
-#region Header
-
-/*
-* MIT License
-
-Copyright (c) 2023 - present DanKe
-
-* DanKeJson
-a simple Json library for the .Net
-JSON : https://json.org
- 
-* About Author
-Name : DanKe
-Address : Guangzhou City , Guangdong Province , China
-Mail : danke1024@foxmail.com
-Github : https://github.com/DanKE123abc
-  */
-
-#endregion
-
 using System.IO;
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace DanKeJson
+namespace DanKeJson.Utils
 {
     /// <summary>
     /// CommentParser
